requirements.txt
flask
python-ldap
mwoauth
pyyaml
toolforge